+++ src/azure-cli/azure/cli/command_modules/alertsmanagement/custom.py
@@ -64,6 +64,9 @@ def cli_alertsmanagement_set_actionrule(client,
                                         suppression_start_time=None,
                                         suppression_end_time=None,
                                         recurrence_values=None):
+    if (action_rule_type not in ['ActionGroup', 'Suppression', 'Diagnostics']) or \
+       (recurrence_type not in ['Always', 'Once', 'Daily', 'Weekly', 'Monthly']):
+        raise CLIError("Invalid input value for --action-rule-type")
 
     action_rule = ActionRule(location="Global", tags={})